--- a/mars/dataframe/arithmetic/core.py
+++ b/mars/dataframe/arithmetic/core.py
@@ -208,11 +208,11 @@ class DataFrameBinOpMixin(DataFrameOperandMixin):
         rhs_is_tensor = isinstance(op.rhs, TENSOR_TYPE)
         tensor, other = (op.rhs, op.lhs) if rhs_is_tensor else (op.lhs, op.rhs)
         if tensor.shape == other.shape:
-            tensor = tensor.rechunk(other.nsplits).single_tiles()
+            tensor = tensor.rechunk(other.nsplits)._inplace_tile()
         else:
             # shape differs only when dataframe add 1-d tensor, we need rechunk on columns axis.
             rechunk_size = other.nsplits[1] if op.axis == 'columns' or op.axis == 1 else other.nsplits[0]
-            tensor = tensor.rechunk((rechunk_size,)).single_tiles()
+            tensor = tensor.rechunk((rechunk_size,))._inplace_tile()
 
         out_chunks = []
         for out_index in itertools.product(*(map(range, other.chunk_shape))):