@qinxuye 2019-11-28
Performance of `tileable.build_graph` is elevated a bit due the reduction of loop in it.

### This PR

```
In [1]: import mars.tensor as mt                                                

In [2]: a = mt.random.rand(1000, 1000, chunk_size=30)                           

In [3]: b = mt.random.rand(1000, 1000, chunk_size=30)                           

In [4]: %time a.dot(b).sum(axis=1).build_graph(tiled=True, compose=False)       
CPU times: user 10.4 s, sys: 138 ms, total: 10.5 s
Wall time: 10.5 s
Out[4]: <mars.graph.DAG at 0xa1af77b90>
```

### Master

```
In [12]: a = mt.random.rand(1000, 1000, chunk_size=30)                          

In [13]: b = mt.random.rand(1000, 1000, chunk_size=30)                          

In [14]: %time a.dot(b).sum(axis=1).build_graph(tiled=True, compose=False)      
CPU times: user 11.3 s, sys: 184 ms, total: 11.4 s
Wall time: 11.4 s
Out[14]: <mars.graph.DAG at 0xa278a3550>
```